requirements.txt
requests
beautifulsoup4
boto3
python-dotenv
apache-airflow-providers-apache-spark